+++ fast_plotter/postproc/__init__.py
@@ -1,0 +1,4 @@
+from .functions import open_many
+
+
+__all__ = ["open_many"]